+++ ios/engine/KMEI/KeymanEngine/Classes/Constants.swift
@@ -26,6 +26,10 @@ public enum Key {
   
   /// Dictionary of lexical model ids keyed by language id in UserDefaults
   static let userPreferredLexicalModels = "UserPreferredLexicalModels"
+  
+  /// Dictionary of prediction/correction toggle settings keyed by language id in UserDefaults
+  static let userPredictSettings = "UserPredictionEnablementSettings"
+  static let userCorrectSettings = "UserCorrectionEnablementSettings"
 
   // Internal user defaults keys
   static let engineVersion = "KeymanEngineVersion"