+++ ios/engine/KMEI/KeymanEngine/Classes/LanguagePicker/KeyboardPickerViewController.swift
@@ -12,8 +12,7 @@ private let toolbarButtonTag = 100
 private let toolbarLabelTag = 101
 private let toolbarActivityIndicatorTag = 102
 
-class KeyboardPickerViewController: UITableViewController, UIAlertViewDelegate {
-  private var userKeyboards: [InstallableKeyboard] = [InstallableKeyboard]()
+class KeyboardPickerViewController: KeyboardSwitcherViewController {
   private var updateQueue: [InstallableKeyboard]?
   private var _isDoneButtonEnabled = false
   private var isDidUpdateCheck = false
@@ -24,11 +23,13 @@ class KeyboardPickerViewController: UITableViewController, UIAlertViewDelegate {
   private var lexicalModelDownloadStartedObserver: NotificationObserver?
   private var lexicalModelDownloadCompletedObserver: NotificationObserver?
   private var lexicalModelDownloadFailedObserver: NotificationObserver?
-
+  
   override func viewDidLoad() {
     super.viewDidLoad()
 
-    title = "Keyboards"
+    self.accessoryType = .detailDisclosureButton
+
+    // adding keyboards is unique to this subclass, so is updating them
     setIsDoneButtonEnabled(false)
     isDidUpdateCheck = false
     updateQueue = nil
@@ -70,13 +71,6 @@ class KeyboardPickerViewController: UITableViewController, UIAlertViewDelegate {
     log.info("didLoad: KeyboardPickerViewController (registered lexicalModelDownloadCompleted et al)")
   }
 
-  override func viewWillAppear(_ animated: Bool) {
-    super.viewWillAppear(animated)
-
-    loadUserKeyboards()
-    scroll(toSelectedKeyboard: false)
-  }
-
   override func viewDidAppear(_ animated: Bool) {
     super.viewDidAppear(animated)
     if isDidUpdateCheck || !checkUpdates() {
@@ -102,54 +96,19 @@ class KeyboardPickerViewController: UITableViewController, UIAlertViewDelegate {
     scroll(toSelectedKeyboard: false)
     log.info("didAppear: KeyboardPickerViewController")
   }
-
-  override func numberOfSections(in tableView: UITableView) -> Int {
-    return 1
+  
+  override func switchKeyboard(_ index: Int) {
+    // Switch keyboard and register to user defaults.
+    if Manager.shared.setKeyboard(userKeyboards[index]) {
+      tableView.reloadData()
+    }
+    
+    if !_isDoneButtonEnabled {
+      Manager.shared.dismissKeyboardPicker(self)
+    }
   }
-
-  override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-    return userKeyboards.count
-  }
-
-  override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-    let cellIdentifier = "Cell"
-    if let cell = tableView.dequeueReusableCell(withIdentifier: cellIdentifier) {
-      return cell
-    }
-
-    let cell = UITableViewCell(style: .subtitle, reuseIdentifier: cellIdentifier)
-    let selectionColor = UIView()
-    selectionColor.backgroundColor = UIColor(red: 204.0 / 255.0, green: 136.0 / 255.0, blue: 34.0 / 255.0, alpha: 1.0)
-    cell.selectedBackgroundView = selectionColor
-    return cell
-  }
-
-  // TODO: Refactor. Duplicated in KeyboardInfoViewController
-  override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
-    if !Manager.shared.canRemoveKeyboards {
-      return false
-    }
-
-    if !Manager.shared.canRemoveDefaultKeyboard {
-      return indexPath.row != 0
-    }
-    if indexPath.row > 0 {
-      return true
-    }
-    return userKeyboards.count > 1
-  }
-
-  override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCellEditingStyle,
-                          forRowAt indexPath: IndexPath) {
-    if editingStyle != .delete {
-      return
-    }
-
-    if Manager.shared.removeKeyboard(at: indexPath.row) {
-      loadUserKeyboards()
-    }
-    setIsDoneButtonEnabled(true)
-  }
+  
+  // MARK: - table view delegate UITableViewDelegate
 
   override func tableView(_ tableView: UITableView,
                           accessoryButtonTappedForRowWith indexPath: IndexPath) {
@@ -171,31 +130,8 @@ class KeyboardPickerViewController: UITableViewController, UIAlertViewDelegate {
     infoView.isCustomKeyboard = kb.isCustom
     navigationController?.pushViewController(infoView, animated: true)
   }
-
-  override func tableView(_ tableView: UITableView,
-                          willDisplay cell: UITableViewCell,
-                          forRowAt indexPath: IndexPath) {
-    cell.selectionStyle = .none
-    let kb = userKeyboards[indexPath.row]
-
-    cell.textLabel?.text = kb.languageName
-    cell.detailTextLabel?.text = kb.name
-    cell.tag = indexPath.row
-
-    if Manager.shared.currentKeyboardID == kb.fullID {
-      cell.selectionStyle = .blue
-      cell.isSelected = true
-      cell.accessoryType = .detailDisclosureButton
-    } else {
-      cell.selectionStyle = .none
-      cell.isSelected = false
-      cell.accessoryType = .detailDisclosureButton
-    }
-  }
-
-  override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-    switchKeyboard(indexPath.row)
-  }
+  
+  // MARK: - keyboard and lexical model downloading
 
   private func keyboardDownloadStarted(_ keyboards: [InstallableKeyboard]) {
     log.info("keyboardDownloadStarted")
@@ -325,26 +261,8 @@ class KeyboardPickerViewController: UITableViewController, UIAlertViewDelegate {
     
     self.present(alertController, animated: true, completion: nil)
   }
-
-  private func switchKeyboard(_ index: Int) {
-    // Switch keyboard and register to user defaults.
-    if Manager.shared.setKeyboard(userKeyboards[index]) {
-      tableView.reloadData()
-    }
-
-    if !_isDoneButtonEnabled {
-      Manager.shared.dismissKeyboardPicker(self)
-    }
-  }
-
-  private func loadUserKeyboards() {
-    userKeyboards = Storage.active.userDefaults.userKeyboards ?? []
-    tableView.reloadData()
-  }
-
-  private func isAdded(languageID langID: String, keyboardID kbID: String) -> Bool {
-    return userKeyboards.contains { kb in kb.languageID == langID && kb.id == kbID }
-  }
+  
+  // MARK: - keyboard adding
 
   @objc func doneClicked(_ sender: Any) {
     Manager.shared.dismissKeyboardPicker(self)
@@ -387,7 +305,8 @@ class KeyboardPickerViewController: UITableViewController, UIAlertViewDelegate {
     updateKeyboards()
   }
 
-  private func checkUpdates() -> Bool {
+  // so KeyboardSwitcherViewController can override this to do nothing
+  public func checkUpdates() -> Bool {
     if Manager.shared.apiKeyboardRepository.languages == nil {
       return false
     }
@@ -419,18 +338,6 @@ class KeyboardPickerViewController: UITableViewController, UIAlertViewDelegate {
     }
   }
 
-  private func scroll(toSelectedKeyboard animated: Bool) {
-    let index = userKeyboards.index { kb in
-      return Manager.shared.currentKeyboardID == kb.fullID
-    }
-
-    if let index = index {
-      let indexPath = IndexPath(row: index, section: 0)
-      tableView.scrollToRow(at: indexPath, at: .middle, animated: animated)
-
-    }
-  }
-
   private func setIsDoneButtonEnabled(_ value: Bool) {
     _isDoneButtonEnabled = value
     if _isDoneButtonEnabled {
@@ -448,7 +355,7 @@ class KeyboardPickerViewController: UITableViewController, UIAlertViewDelegate {
     navigationController?.setToolbarHidden(true, animated: true)
   }
 
-  func showAddKeyboard() {
+  override func showAddKeyboard() {
     let button: UIButton? = (navigationController?.toolbar?.viewWithTag(toolbarButtonTag) as? UIButton)
     button?.isEnabled = false
     let vc = LanguageViewController(Manager.shared.apiKeyboardRepository)