+++ ios/engine/KMEI/KeymanEngine/Classes/LanguagePicker/LanguageDetailViewController.swift
@@ -50,8 +50,10 @@ class LanguageDetailViewController: UITableViewController, UIAlertViewDelegate {
   override open func viewWillAppear(_ animated: Bool) {
     super.viewWillAppear(animated)
     
-    log.info("didAppear: LanguageDetailViewController (actually willAppear)")
+    log.info("willAppear: LanguageDetailViewController")
   }
+
+  // MARK: - Table view data source UITableViewDataSource
 
   override func numberOfSections(in tableView: UITableView) -> Int {
     return language.keyboards!.count