+++ ios/engine/KMEI/KeymanEngine/Classes/Settings/LanguageSettingsViewController.swift
@@ -78,8 +78,38 @@ class LanguageSettingsViewController: UITableViewController {
                              height: cellFrame.size.height)
     return switchFrame
   }
+  
+  @objc
+  func predictionSwitchValueChanged(source: UISwitch) {
+    let value = source.isOn;
+    let userDefaults = Storage.active.userDefaults
+    userDefaults.set(predictSetting: value, forLanguageID: self.language.id)
+    
+    if let lm = Manager.shared.preferredLexicalModel(userDefaults, forLanguage: self.language.id) {
+      if Manager.shared.currentKeyboardID?.languageID == self.language.id {
+        // re-register the model - that'll enact the settings.
+        _ = Manager.shared.registerLexicalModel(lm)
+      }
+    }
+  }
+  
+  @objc
+  func correctionSwitchValueChanged(source: UISwitch) {
+    let value = source.isOn;
+    let userDefaults = Storage.active.userDefaults
+    userDefaults.set(correctSetting: value, forLanguageID: self.language.id)
+    
+    if let lm = Manager.shared.preferredLexicalModel(userDefaults, forLanguage: self.language.id) {
+      if Manager.shared.currentKeyboardID?.languageID == self.language.id {
+        // re-register the model - that'll enact the settings.
+        _ = Manager.shared.registerLexicalModel(lm)
+      }
+    }
+  }
 
   override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+    let userDefaults = Storage.active.userDefaults
+    
     let cellIdentifier = 0 == indexPath.section ?  "KeyboardInLanguageSettingsCell" : "LanguageSettingsCell"
     
     let reusableCell = tableView.dequeueReusableCell(withIdentifier: cellIdentifier)
@@ -89,23 +119,36 @@ class LanguageSettingsViewController: UITableViewController {
     
     let cell = UITableViewCell(style: .subtitle, reuseIdentifier: cellIdentifier)
     if 1 == indexPath.section {
-        //TODO: find the settings these are to show
         if 0 == indexPath.row {
           cell.accessoryType = .none
           let doPredictionsSwitch = UISwitch()
+          doPredictionsSwitch.translatesAutoresizingMaskIntoConstraints = false
+          
           let switchFrame = frameAtRightOfCell(cell: cell.frame, controlSize: doPredictionsSwitch.frame.size)
           doPredictionsSwitch.frame = switchFrame
-          doPredictionsSwitch.isOn = false
-          //            showBannerSwitch.addTarget(self, action: #selector(self.predictionSwitchValueChanged), for: .valueChanged)
+          
+          doPredictionsSwitch.isOn = userDefaults.predictSettingForLanguage(languageID: self.language.id)
+          doPredictionsSwitch.addTarget(self, action: #selector(self.predictionSwitchValueChanged), for: .valueChanged)
           cell.addSubview(doPredictionsSwitch)
+          if #available(iOSApplicationExtension 9.0, *) {
+            doPredictionsSwitch.rightAnchor.constraint(equalTo: cell.layoutMarginsGuide.rightAnchor).isActive = true
+            doPredictionsSwitch.centerYAnchor.constraint(equalTo: cell.layoutMarginsGuide.centerYAnchor).isActive = true
+          }
         } else if 1 == indexPath.row {
           cell.accessoryType = .none
           let doCorrectionsSwitch = UISwitch()
+          doCorrectionsSwitch.translatesAutoresizingMaskIntoConstraints = false
+          
           let switchFrame = frameAtRightOfCell(cell: cell.frame, controlSize: doCorrectionsSwitch.frame.size)
           doCorrectionsSwitch.frame = switchFrame
-          doCorrectionsSwitch.isOn = false
-          //            showBannerSwitch.addTarget(self, action: #selector(self.correctionSwitchValueChanged), for: .valueChanged)
+          
+          doCorrectionsSwitch.isOn = userDefaults.correctSettingForLanguage(languageID: self.language.id)
+          doCorrectionsSwitch.addTarget(self, action: #selector(self.correctionSwitchValueChanged), for: .valueChanged)
           cell.addSubview(doCorrectionsSwitch)
+          if #available(iOSApplicationExtension 9.0, *) {
+            doCorrectionsSwitch.rightAnchor.constraint(equalTo: cell.layoutMarginsGuide.rightAnchor).isActive = true
+            doCorrectionsSwitch.centerYAnchor.constraint(equalTo: cell.layoutMarginsGuide.centerYAnchor).isActive = true
+          }
         } else { // rows 3 and 4
           cell.accessoryType = .disclosureIndicator
       }
@@ -124,9 +167,9 @@ class LanguageSettingsViewController: UITableViewController {
     var title: String
     switch (section) {
     case 0:
-      title = "keyboards"
+      title = "Keyboards"
     case 1:
-      title = "language settings"
+      title = "Language settings"
     default:
       title = "unknown header"
     }
@@ -166,23 +209,23 @@ class LanguageSettingsViewController: UITableViewController {
       cell.accessoryType = .none
       switch indexPath.row {
         case 0:
-          cell.textLabel?.text = "Enable corrections"
-        case 1:
           cell.textLabel?.text = "Enable predictions"
+        case 1:
+          cell.textLabel?.text = "Enable corrections"
         case 2:
-          cell.textLabel?.text = "Model"
+          cell.textLabel?.text = "Dictionaries"
           cell.accessoryType = .disclosureIndicator
           if let modelCt = language.lexicalModels?.count {
             switch modelCt {
             case 0:
-              cell.detailTextLabel?.text = "no models installed"
+              cell.detailTextLabel?.text = "no dictionaries installed"
             case 1:
-              cell.detailTextLabel?.text = "one model installed"
+              cell.detailTextLabel?.text = "\(language.lexicalModels![0].name)"
             default:
-              cell.detailTextLabel?.text = "\(modelCt) models installed"
+              cell.detailTextLabel?.text = "\(modelCt) dictionaries installed"
             }
           } else {
-            cell.detailTextLabel?.text = "no models installed"
+            cell.detailTextLabel?.text = "no dictionaries installed"
           }
         case 3: // future
           cell.textLabel?.text = "Manage dictionary"
@@ -204,6 +247,40 @@ class LanguageSettingsViewController: UITableViewController {
     performAction(for: indexPath)
   }
   
+  override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
+    if !Manager.shared.canRemoveKeyboards {
+      return false
+    }
+    
+    if indexPath.section != 0 {
+      return false
+    }
+    
+    // Filter- prevent deleting the default keyboard and just that one.
+    if let globalIndex = getKeyboardIndex(kb: (language.keyboards?[safe: indexPath.row])!) {
+      // Assumption - default keyboard is index 0.  Probably should make something more robust, though.
+      if globalIndex == 0 {
+        return false
+      }
+    }
+
+    return true
+  }
+  
+  override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCellEditingStyle,
+                          forRowAt indexPath: IndexPath) {
+    if editingStyle == .delete {
+      if let globalIndex = getKeyboardIndex(kb: (language.keyboards?[safe: indexPath.row])!) {
+        if Manager.shared.removeKeyboard(at: globalIndex) {
+          // For now, a pop-back will be sufficient.
+          navigationController?.popToRootViewController(animated: true)
+        }
+      }
+    }
+
+    // Do nothing for now.
+  }
+  
   @objc func addClicked(_ sender: Any) {
     showAddLanguageKeyboard()
   }
@@ -222,6 +299,7 @@ class LanguageSettingsViewController: UITableViewController {
       showKeyboardInfoView(kb: (language.keyboards?[safe: indexPath.row])!)
     case 1:
       switch indexPath.row  {
+        // case 0, 1:  the toggles - but a general 'click' not on the toggle itself.
         case 2:
           showLexicalModelsView()
         default:
@@ -229,6 +307,27 @@ class LanguageSettingsViewController: UITableViewController {
       }
     default:
       break
+    }
+  }
+  
+  func getKeyboardIndex(kb: Keyboard) -> Int? {
+    let matchingFullID = FullKeyboardID(keyboardID: kb.id, languageID: language.id)
+    let userData = Storage.active.userDefaults
+
+    // If user defaults for keyboards list does not exist, do nothing.
+    guard var globalUserKeyboards = userData.userKeyboards else {
+      log.error("no keyboards in the global keyboards list!")
+      return nil
+    }
+
+    if let index = globalUserKeyboards.index(where: { $0.fullID == matchingFullID }) {
+      guard index < globalUserKeyboards.count else {
+        return nil
+      }
+      return index
+    } else {
+      log.error("this keyboard \(matchingFullID) not found among user's installed keyboards!")
+      return nil
     }
   }
   
@@ -244,7 +343,7 @@ class LanguageSettingsViewController: UITableViewController {
       return
     }
 
-    if let index = globalUserKeyboards.index(where: { $0.fullID == matchingFullID }) {
+    if let index = getKeyboardIndex(kb: kb) {
       guard index < globalUserKeyboards.count else {
         return
       }
@@ -267,7 +366,7 @@ class LanguageSettingsViewController: UITableViewController {
   
   func showLexicalModelsView() {
     //LanguageLexicalModelPickerViewController? (should show just the models for this language)
-    let lmListView = LexicalModelPickerViewController()
+    let lmListView = LexicalModelPickerViewController(self.language)
     lmListView.language = self.language
     navigationController?.pushViewController(lmListView, animated: true)
  }